+++ src/commands/CopyConnectorIdCommand.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+import { ConnectorTreeItem } from "../models/TreeModel";
+
+export class CopyConnectorIdCommand {
+
+    public async execute(item: ConnectorTreeItem) {
+        vscode.env.clipboard.writeText(item.id);
+    }
+}